externalize ginkgo_exec + remove cmake options
