+++ benchmarks/CMakeLists.txt
@@ -1,7 +1,5 @@
 # SPDX-License-Identifier: MIT
 
-option(BUILD_SPLINES_KERNEL  "Build DDC kernel for splines" ON)
-
 add_executable(ddc_benchmark_deepcopy deepcopy.cpp)
 target_link_libraries(ddc_benchmark_deepcopy
 	PUBLIC
